working db actions
